Update README with new prerequisites and production usage instructions; refactor UI components to a dedicated folder and enhance button styles for consistency.
