+++ app/layout.tsx
@@ -3,7 +3,7 @@ import { Geist, Geist_Mono } from "next/font/google";
 import { Analytics } from "@vercel/analytics/react";
 import "./globals.css";
 import "@rainbow-me/rainbowkit/styles.css";
-import { DatadogAnalytics } from "@/components/datadog-analytics";
+import { DatadogAnalytics } from "@/components/UI/datadog-analytics";
 import { Providers } from "@/app/providers";
 
 const geistSans = Geist({
